Added Unholiday file to contain the days which were supposed to be holidays, but in which you got stuck at work.
The unholiday file has the same format as the holiday file.
Notice: The unholiday file has precedence over the holiday file: if a
date is inserted in both, then it counts as a working day.

Notice: The unholiday file was also put in the .gitignore list; there is
no reason to be sharing out own unholidays between us.